Hold the repo to the standard it argues for: lint, types, and a gate that can fail
The project's claim is that a tool is only worth trusting when something
independent can fail it. That applied to everything except this repo, which had
tests and nothing else. Now CI enforces six things on every pull request.

- ruff and ruff format across src, tests, examples and scripts. BLE is in the
  ruleset deliberately: every broad `except` in this package now carries a
  written justification next to it, and deleting the reason brings the lint
  back.
- mypy, clean across all 18 modules, and the package ships py.typed so an
  adopting project type-checks against it rather than against Any.
- Coverage floor of 75% with branch coverage on (currently 78%). Not 90: the
  PostgreSQL source cannot run without a live server, and mocking a database to
  raise a percentage would only test the mock.
- The stdio smoke test moved into CI, so a real client spawning a real
  subprocess is proven on every push rather than by hand.
- `lint broken_checkout` must exit 1. A lint that cannot fail is decorative,
  and the assertion belongs in CI rather than in a reviewer's memory.

Also fixes a genuine hole found by the type checker: Threshold.band() fell
through to formatting a None when a threshold had neither bound. Unreachable
today because __post_init__ rejects that case, but it was relying on a check
three methods away, and now it raises with its own message.

Adds SECURITY.md (the read-only boundary, the redaction guarantee, and what is
explicitly out of scope — your own toolkit code runs with your privileges),
CHANGELOG.md, issue and PR templates, and dependabot for the workflow actions.

+++ src/groundtruth_mcp/render.py
@@ -15,7 +15,7 @@ of the conversation. Three rules hold throughout:
 from __future__ import annotations
 
 import json
-from typing import Iterable, Sequence
+from collections.abc import Iterable, Sequence
 
 from .budget import fence_untrusted, truncate
 from .contracts import Report, Trace
@@ -153,7 +153,7 @@ def render_targets(targets: dict[str, str], noun: str, *, limit: int) -> str:
 def _number(value: float) -> str:
     """Two decimals for fractions, no decimals for whole counts."""
     if abs(value - round(value)) < 1e-9:
-        return str(int(round(value)))
+        return str(round(value))
     return f"{value:.2f}"
